Много вся кого

+++ TaskManager.Api/Controllers/ProjectsController.cs
@@ -2,19 +2,29 @@
 using Microsoft.AspNetCore.Mvc;
 using TaskManager.Api.Controllers.Abstracted;
 using TaskManager.Api.Data;
+using TaskManager.Api.Entity;
 using TaskManager.Api.Services;
 using TaskManager.Command.Models;
+using Microsoft.AspNetCore.Http;
+using System.Data;
 
 namespace TaskManager.Api.Controllers
 {
     [Authorize]
     [ApiController]
     [Route("api/[controller]")]
-    public class ProjectsController : ControllerBase, ICRUDController<ProjectModel>
+    public class ProjectsController : ControllerBase
     {
         private readonly ProjectService _service;
+        private readonly RoleService _roleService;
+        private readonly HttpContextHandlerService _httpHandler;
 
-        public ProjectsController(ApplicationContext context) { _service = new(context); }
+        public ProjectsController(ApplicationContext context) 
+        { 
+            _service = new(context);
+            _httpHandler = new(context);
+            _roleService = new(context);
+        }
 
         /// <summary>
         /// Create a project
@@ -26,8 +36,10 @@ namespace TaskManager.Api.Controllers
         [ProducesResponseType(typeof(Response<ProjectModel>), StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Create([FromBody] ProjectModel model)
         {
-            if (model is null)
-                return BadRequest(new Response<ProjectModel> { IsSuccess = false, Reason = "Request null" });
+            var user = _httpHandler.GetUserAsNoTracking(HttpContext);
+            if (user == null)
+                return BadRequest();
+            model.CreatorId = user.Id;
             var modelToCreate = await _service.CreateAsync(model);
             if (!modelToCreate.IsSuccess)
                 return BadRequest(modelToCreate);
@@ -44,26 +56,13 @@ namespace TaskManager.Api.Controllers
         [ProducesResponseType(typeof(Response), StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!IsValidate(_httpHandler.GetUserRoleAsNoTracking(HttpContext, id)?.AllowedDeleteProject))
+                return BadRequest(new Response { IsSuccess = false, Reason = "no access" });
             var project = await _service.GetByIdAsync(id);
             if (!project.IsSuccess)
                 return BadRequest(project);
             _service.Delete(id);
             return Ok(project);
-        }
-
-        /// <summary>
-        /// Get all projects
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet]
-        [ProducesResponseType(typeof(Response<List<ProjectModel>>), StatusCodes.Status200OK)]
-        [ProducesResponseType(typeof(Response<List<ProjectModel>>), StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> GetAll()
-        {
-            var response = await _service.GetAllAsync();
-            if (!response.IsSuccess)
-                return BadRequest(response);
-            return Ok(response);
         }
 
         /// <summary>
@@ -83,7 +82,7 @@ namespace TaskManager.Api.Controllers
         }
 
         /// <summary>
-        /// Update the project
+        /// Edit the project
         /// </summary>
         /// <param name="id"></param>
         /// <param name="model"></param>
@@ -91,8 +90,10 @@ namespace TaskManager.Api.Controllers
         [HttpPut]
         [ProducesResponseType(typeof(Response), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(Response), StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> Update(int id, ProjectModel model)
+        public async Task<IActionResult> Edit(int id, ProjectModel model)
         {
+            if (!IsValidate(_httpHandler.GetUserRoleAsNoTracking(HttpContext, id)?.AllowedEditProject))
+                return BadRequest(new Response { IsSuccess = false, Reason = "no access" });
             var response = await _service.UpdateAsync(model);
             if (!response.IsSuccess)
                 return BadRequest(response);
@@ -100,16 +101,36 @@ namespace TaskManager.Api.Controllers
         }
 
 
-        //[HttpPost("/{projectId}/Users/{userId}")]
-        //public Task<IActionResult> AddUser(int projectId, int userId)
-        //{
-        //    ProjectParticipant participant = new()
-        //    {
-        //        UserId = userId,
-        //        ProjectId = projectId,
-        //    };
-        //    _participantService.Create(participant);
-        //    return Ok();
-        //}
+        [HttpPost("/{projectId}/Users")]
+        [ProducesResponseType(typeof(Response), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(Response), StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> AddUsers(int projectId, int[] usersId)
+        {
+            var response = await _service.AddUsers(projectId, usersId);
+            if (!response.IsSuccess)
+                return BadRequest(response);
+            return Ok(response);
+        }
+
+        [HttpGet("/{id}/Users")]
+        [ProducesResponseType(typeof(Response<List<UserRoleModel>>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(Response<List<UserRoleModel>>), StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> GetUsers(int id)
+        {
+            var role = _httpHandler.GetUserRoleAsNoTracking(HttpContext, id);
+            if (role is null)
+                return BadRequest(new Response { IsSuccess = false, Reason = "no access" });
+            var response = await _service.GetUsers(id);
+            if (!response.IsSuccess)
+                return BadRequest(response);
+            return Ok(response);
+        }
+
+        private bool IsValidate(bool? isValidate)
+        {
+            if (isValidate is null)
+                return false;
+            return (bool)isValidate;
+        }
     }
 }
